refactor: Move package of vaildStringExpressionChecker from checker to expression
