+++ src/test/java/expression/ValidStringExpressionCheckerTests.java
@@ -1,13 +1,12 @@
-package checker;
+package expression;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Arrays;
+import expression.ValidStringExpressionChecker;
 
 public class ValidStringExpressionCheckerTests {
